pyproject.toml
[project]
name = "e-shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
]
